Break catalog/registrations circular import by owning catalog data in registrations.
Alignment helpers now take an explicit catalog tuple so catalog.py stays types-only and BFF callers pass TURN_ANALYTIC_CATALOG directly.
