+++ packages/api/api/analytics/__init__.py
@@ -1,6 +1,6 @@
 """Core turn analytics."""
 
-from api.analytics.catalog import TURN_ANALYTIC_CATALOG, TurnAnalyticCatalogEntry
+from api.analytics.catalog import TurnAnalyticCatalogEntry
 from api.analytics.compute_context import AnalyticComputeContext
 from api.analytics.options import TurnAnalyticsOptions
 from api.analytics.registration import (
@@ -10,7 +10,11 @@ from api.analytics.registration import (
     handler_from_turn,
     handler_from_turn_and_options,
 )
-from api.analytics.registrations import TURN_ANALYTIC_REGISTRATIONS
+from api.analytics.registrations import (
+    TURN_ANALYTIC_CATALOG,
+    TURN_ANALYTIC_REGISTRATIONS,
+    catalog_entry,
+)
 from api.analytics.registry import TURN_ANALYTICS, get_turn_analytic
 
 __all__ = [
@@ -18,6 +22,7 @@ __all__ = [
     "EmptyExportCatalog",
     "handler_from_turn",
     "handler_from_turn_and_options",
+    "catalog_entry",
     "TURN_ANALYTIC_CATALOG",
     "TURN_ANALYTIC_REGISTRATIONS",
     "TURN_ANALYTICS",
