+++ packages/bff/bff/analytics/scores.py
@@ -1,6 +1,6 @@
 """BFF Scores table analytic handler."""
 
-from api.analytics.catalog import catalog_entry
+from api.analytics.registrations import catalog_entry
 from api.diagnostics import Diagnostics
 
 from bff.analytics.descriptor import AnalyticDescriptor
